feat(core): the variants axis carries its number type (#556 PR 1b-i)
`number_operator_contract!`'s `numbers:` axis was inert. The declared type
reached exactly one expression — the struct-name fragment — and `to_spec`
hardcoded `F32Meta`/`PortTy::F32`/`F32Buffer` throughout. `numbers: [i32]`
would have emitted an operator *named* `add_i32_value` whose every port was
`f32`: a well-formed contract, wrong on the wire, caught by nothing.

Replace the two axes with a written `variants:` list, one `<number type>
<carrier>` entry per operator, and carry the type through to the ports.

The list is not a `numbers × carriers` product because the product is not
full — `i32` has no dense buffer form, so `i32 signal` is rejected at the
parse rather than needing a per-number carrier table to skip a cell. It also
leaves room for the converter operators, whose input and output types differ.

One operand declaration serves every entry: bounds and defaults are parsed
into a type-neutral `f64` and projected per variant, so `default 1` is `1.0`
at f32 and `1` at i32. A value that cannot survive the projection — a
fractional default on an op that also lists `i32` — is a spanned compile
error, not an `as i32` truncation. `NUMBER_MIN_I32`/`NUMBER_MAX_I32` are
derived from their f32 counterparts rather than restated, so the type-wide
span cannot drift between the two types.

Ships the 10 i32 math operators the axis now makes expressible: add sub mul
div min max abs negate modulo clamp. `power` (#561) and `map` (#562) need a
different signature and a different algorithm; `reciprocal` compiles at i32
and is useless there (`1/n == 0` for all |n| > 1). Each omission is recorded
in the module doc, because only the first is a compile error.

Overflow was the gap the plan of record did not anticipate. `i32`'s `+` and
`*` panic on overflow in a debug build and wrap in a release one, where f32
yields `inf` — and `process` runs on the render thread, where a panic is
fatal. The declared port range does not save it: every inbound value is
clamped, but two operands at the type-wide ±1e6 sentinel multiply to 1e12,
past `i32::MAX`. So the five operations that can leave the range bind a new
`PointwiseNum` (add/sub/mul/neg/abs) that saturates at each type's limits,
making `inf` and `i32::MAX` the same answer in two types. num-traits'
saturating traits are integer-only and cannot express the f32 half.

Verification: 997 tests pass, 0 fail (974 on dev; the +23 is 5 macro tests,
2 `PointwiseNum` tests, 10 generated `defaults_are_data` tests, and 6
hand-written i32 arithmetic tests). fmt, clippy, and the three
`check_rules_*.py` guards are clean, and rustdoc warnings hold at dev's 81.

`reuben describe --json` is **purely additive** against dev: 57 -> 67
operators, zero removed, zero changed, the 10 added exactly as intended. No
instrument document migrates.

Docs corrected here: the rule rationale's two claims the issue flags are now
true, and `power.rs` gave the wrong reason for its f32-only restriction — it
blamed `max` and `powf` jointly, but `x.max(0)` is `PartialOrd` + a zero and
generalizes trivially. The real obstacle is heterogeneous operand types.

Refs #556. Follow-ups: #560 · #561 · #562 · #565.

+++ crates/reuben-macros/src/grammar.rs
@@ -1,27 +1,33 @@
 //! Shared numeric-literal parsing for both contract macros (`operator_contract!` and
 //! `number_operator_contract!`).
 //!
-//! A `f32` range endpoint or `default` may be written as a literal *or* as the `min`/`max` sentinel
+//! A range endpoint or `default` may be written as a literal *or* as the `min`/`max` sentinel
 //! keyword (issue #127), so the `±1e6` type-wide range never appears as a raw literal in an operator
 //! contract. In a **range endpoint** the sentinel resolves to the shared
 //! [`NUMBER_MIN`]/[`NUMBER_MAX`] type-wide bound; in a **`default`** it resolves to the operand's own
 //! declared range extreme (already parsed), so `default max` parks an operand at its ceiling
 //! regardless of what that ceiling is.
+//!
+//! Everything here answers in `f64` — the **type-neutral** form, not the port's type. A literal is
+//! parsed once and projected by the caller onto whatever type the port turned out to be: `f32` for
+//! `parse_f32_meta`, and per-variant for `number_operator_contract!`, whose one operand declaration
+//! serves instantiations at more than one number type (issue #556). `f64` holds every `f32` literal
+//! and every `i32` exactly, so the projection is the only place precision is ever at stake.
 
 use reuben_contract::{NUMBER_MAX, NUMBER_MIN};
 use syn::parse::ParseStream;
 use syn::{Error, Ident, Lit, Token};
 
 /// A numeric literal with an optional leading `-` (bounds and defaults can be negative).
-pub(crate) fn parse_signed_float(input: ParseStream) -> syn::Result<f32> {
+pub(crate) fn parse_signed_float(input: ParseStream) -> syn::Result<f64> {
     let neg = input.peek(Token![-]);
     if neg {
         input.parse::<Token![-]>()?;
     }
     let lit: Lit = input.parse()?;
     let val = match lit {
-        Lit::Float(f) => f.base10_parse::<f32>()?,
-        Lit::Int(i) => i.base10_parse::<f32>()?,
+        Lit::Float(f) => f.base10_parse::<f64>()?,
+        Lit::Int(i) => i.base10_parse::<f64>()?,
         other => return Err(Error::new(other.span(), "expected a numeric literal")),
     };
     Ok(if neg { -val } else { val })
@@ -37,12 +43,12 @@ pub(crate) fn peek_range_sentinel(input: ParseStream) -> bool {
 
 /// A range endpoint: a signed literal, or the `min`/`max` sentinel resolving to the type-wide
 /// [`NUMBER_MIN`]/[`NUMBER_MAX`] bound (issue #127).
-pub(crate) fn parse_float_or_sentinel(input: ParseStream) -> syn::Result<f32> {
+pub(crate) fn parse_float_or_sentinel(input: ParseStream) -> syn::Result<f64> {
     if input.peek(Ident) {
         let id: Ident = input.parse()?;
         return match id.to_string().as_str() {
-            "min" => Ok(NUMBER_MIN),
-            "max" => Ok(NUMBER_MAX),
+            "min" => Ok(f64::from(NUMBER_MIN)),
+            "max" => Ok(f64::from(NUMBER_MAX)),
             other => Err(Error::new(
                 id.span(),
                 format!("expected a number or the `min`/`max` sentinel, got `{other}`"),
@@ -55,7 +61,7 @@ pub(crate) fn parse_float_or_sentinel(input: ParseStream) -> syn::Result<f32> {
 /// A `default` value: a signed literal, or `max`/`min` resolving to the operand's **own** range
 /// extreme (`hi`/`lo` — the endpoints parsed for this same operand). Parks an operand at its ceiling
 /// (`min`'s no-op `b`) or floor (`max`'s) without repeating the bound (issue #127).
-pub(crate) fn parse_default_value(input: ParseStream, lo: f32, hi: f32) -> syn::Result<f32> {
+pub(crate) fn parse_default_value(input: ParseStream, lo: f64, hi: f64) -> syn::Result<f64> {
     if input.peek(Ident) {
         let id: Ident = input.parse()?;
         return match id.to_string().as_str() {
